Implement authentication and authorization features with Spring Security

+++ src/main/java/com/url_service/url_service/repository/UrlRepository.java
@@ -3,5 +3,5 @@ package com.url_service.url_service.repository;
 import com.url_service.url_service.models.Url;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UrlRepository extends JpaRepository<Url, Integer> {
+public interface UrlRepository extends JpaRepository<Url, Long> {
 }
